Cleaning up Code that Wasn't Used
Removed a copy of the SiteResolver.cs from Sitecore that was sitting in the solution that I used for debug purposes to ensure that my module was working as intended. However, I am removing this in relation to a Stack Exchange answer that I'm providing about this module as to why I don't use the httpBeginRequest pipeline, and didn't want anyone coming in here thinking I did use it, because it was in the solution. So there. =P